+++ src/main/java/com/dev_incubator/dits/service/interfaces/AnswerServiceY.java
@@ -3,5 +3,5 @@ package com.dev_incubator.dits.service.interfaces;
 import com.dev_incubator.dits.persistence.entity.AnswerY;
 
 public interface AnswerServiceY {
-    AnswerY getCorrectByDescription(int id);
+    AnswerY getCorrectByDescription(Long id);
 }
