+++ src/main/java/com/dev_incubator/dits/service/interfaces/UserServiceY.java
@@ -3,5 +3,5 @@ package com.dev_incubator.dits.service.interfaces;
 import com.dev_incubator.dits.persistence.entity.UserY;
 
 public interface UserServiceY {
-    UserY findUserById(int id);
+    UserY findUserById(Long id);
 }
